% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{controlmcmcSDA}
\alias{controlmcmcSDA}
\title{MCMC Control for SDALGCP}
\usage{
controlmcmcSDA(n.sim, burnin, thin, h, c1.h, c2.h)
}
\arguments{
\item{n.sim}{total number of simulations.}

\item{burnin}{number of burn-in iterations.}

\item{thin}{thinning interval.}

\item{h}{tuning parameter (defaults internally if missing).}

\item{c1.h}{adaptation control parameter.}

\item{c2.h}{adaptation control parameter.}
}
\value{
A named list for use in \code{\link{SDALGCPMCML}}.
}
\description{
Specifies MCMC sampling settings for the Langevin-Hastings algorithm.
}
\examples{
h <- 1.65 / (545^(1/6))
control <- controlmcmcSDA(n.sim = 10000, burnin = 2000, thin = 8, h = h, c1.h = 0.01, c2.h = 1e-4)
str(control)
}
