% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBCshp.R
\docType{data}
\name{PBCshp}
\alias{PBCshp}
\title{PBC count data and index of multiple deprivation data.}
\format{A SpatialPolygonsDataFrame of object containing the PBC cases count for each LSOA in Newcastle upon Tyne, UK, as well as the index of multiple deprivation.
\describe{
  \item{X}{PBC count}
  \item{pop}{population count}
  \item{LSOA04CD}{LSOA ID}
  \item{pop}{population count}
  \item{males}{number of males}
  \item{females}{number of females}
  \item{propmale}{proportion of males}
  \item{IMD}{index of multiple deprivation score}
  \item{Income}{proportion of the population experiencing income deprivation}
  \item{Employment}{proportion of the population experiencing employment deprivation}
  \item{Health}{deprivation due to Health}
  \item{Education}{deprivation due to education}
  \item{Barriers}{arriers to housing and services}
  \item{Crime}{deprivation due to crime}
  \item{Environment}{living environment deprivation}
  ...
}}
\usage{
data(PBCshp)
}
\description{
A dataset containing PBC count and Index of multiple deprivation
}
\keyword{datasets}
