% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{plot.Pred.SDALGCPST}
\alias{plot.Pred.SDALGCPST}
\title{plot.Pred.SDALGCPST function}
\usage{
\method{plot}{Pred.SDALGCPST}(
  x,
  type = "relrisk",
  continuous = NULL,
  thresholds = NULL,
  bound = NULL,
  overlay = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of class "Pred.SDALGCPST" obtained as result of a call to \code{\link{SDALGCPPred_ST}}.}

\item{type}{Character string: what type of plot to produce. For discrete inference choices are "incidence" (=exp(mu+S)); "SEincidence" (standard error of incidence); "CovAdjRelRisk" (=exp(S)); or "SECovAdjRelRisk" (standard error of covariate adjusted relative risk); while for continuous inference, choices are "relrisk" (=exp(S)); "SErelrisk" (standard error of the relative risk).}

\item{continuous}{logical; TRUE for spatially continuous relative risk and FALSE for region specific relative risk. default is TRUE}

\item{thresholds}{optional; (only used if you want to plot the exceedance probability) either a vector  of numbers or a vector of single value.}

\item{bound}{optional; it gives the boundary of the region, only useful when the predictive location is supplied in \link{SDALGCPPred_ST}}

\item{overlay}{optional; a logical operation to indicate either to add a base map.}

\item{...}{further arguments passed to \link{plot}.}
}
\value{
The function does not return any value.
}
\description{
Simple plotting function for both discrete and continuous prediction from the object of class "Pred.SDALGCPST".
}
\details{
This function plots the inference from \code{\link{SDALGCPPred}} function. It plots for region-specific inference; incidence and covariate adjusted relative risk while for spatially continuous inference it plots the relative risk. It can as well plot the exceedance probability for spatially discrete and continuous inference.
}
\examples{
# check vignette for examples
}
\seealso{
\link{SDALGCPPred_ST}, \link{plot_continuousST}, \link{plot_discreteST}, \link{plot_SDALGCPexceedanceST}, \link{SDALGCPexceedanceST}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
