% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{plot_discrete}
\alias{plot_discrete}
\title{Plot of the spatially discrete prediction over the region}
\usage{
plot_discrete(obj, type = "incidence", ...)
}
\arguments{
\item{obj}{an object of class "Pred.SDALGCP" obtained as result of a call to \code{\link{SDALGCPPred}}}

\item{type}{Character string: what type of plot to produce. Choices are "incidence" (=exp(mu+S)); "SEincidence" (standard error of incidence); "CovAdjRelRisk" (=exp(S)); or "SECovAdjRelRisk" (standard error of covariate adjusted relative risk);.}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\value{
The function does not return any value.
}
\description{
This function plots the region specific incidence and covariate adjusted relative risk and their corresponding standard error using the output from \code{\link{SDALGCPPred}} function.
}
\seealso{
\code{\link{SDALGCPPred}}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
