% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{plot_discreteST}
\alias{plot_discreteST}
\title{plot_discrete}
\usage{
plot_discreteST(obj, type = "incidence", overlay = FALSE, ...)
}
\arguments{
\item{obj}{an object of class "Pred.SDALGCPST" obtained as result of a call to \code{\link{SDALGCPPred_ST}}}

\item{type}{Character string: what type of plot to produce. Choices are "incidence" (=exp(mu+S)); "SEincidence" (standard error of incidence); "CovAdjRelRisk" (=exp(S)); or "SECovAdjRelRisk" (standard error of covariate adjusted relative risk);.}

\item{overlay}{optional; a logical operation to indicate either to add a base map.}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\value{
The function does not return any value.
}
\description{
A generic function for mapping spatially discrete prediction for \code{\link{SDALGCPPred_ST}} function in \link{SDALGCP} package. Not for general purposes
}
\seealso{
\code{\link{SDALGCPPred}}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
\keyword{internal}
