% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{print.summary.SDALGCP}
\alias{print.summary.SDALGCP}
\title{Print method for summary of SDALGCP model}
\usage{
\method{print}{summary.SDALGCP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.SDALGCP}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Prints the summary of parameter estimates from an SDALGCP model.
}
