% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDALGCP-package.R
\docType{package}
\name{SDALGCP-package}
\alias{SDALGCP}
\alias{SDALGCP-package}
\title{SDALGCP: A package to make continuous inference from spatially aggregated disease count data.}
\description{
The SDALGCP package provides four main functions:
\code{\link{SDALGCPMCML}}, \code{\link{SDALGCPMCML_ST}}, \code{\link{SDALGCPPred}}, and \code{\link{SDALGCPPred_ST}}.
}
\section{SDALGCP functions}{

The \link{SDALGCPMCML} function uses Monte Carlo Maximum Likelihood to estimate the parameters of a
Poisson log-linear model with a spatially continuous random effect for static spatial datasets.

The \link{SDALGCPPred} function delivers spatially discrete prediction of the incidence and the
covariate-adjusted relative risk, and spatially continuous prediction of the covariate-adjusted relative risk.

The \link{SDALGCPMCML_ST} function uses Monte Carlo Maximum Likelihood to estimate the parameters of a
Poisson log-linear model with a spatially continuous random effect for spatio-temporal datasets.

The \link{SDALGCPPred_ST} function provides predictions for both spatially discrete and continuous risk
in the spatio-temporal setting.

Functions such as \link{summary}, \link{confint}, and \link{print} can also be applied to the model outputs.
}

\references{
Christensen, O. F. (2004). Monte Carlo maximum likelihood in model-based geostatistics.
*Journal of Computational and Graphical Statistics*, 13(3), 702–718. \doi{10.1198/106186004X2525}

Giorgi, E., & Diggle, P. J. (2017). PrevMap: an R package for prevalence mapping.
*Journal of Statistical Software*, 78(8), 1–29. \doi{10.18637/jss.v078.i08}

Banerjee, S., Carlin, B. P., & Gelfand, A. E. (2014). *Hierarchical Modeling and Analysis for Spatial Data*. CRC Press.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/olatunjijohnson/SDALGCP}
  \item Report bugs at \url{https://github.com/olatunjijohnson/SDALGCP/issues}
}

}
\author{
Olatunji O. Johnson, Emanuele Giorgi, Peter Diggle. All from CHICAS, Lancaster Medical School,
Faculty of Health and Medicine, Lancaster University
}
\keyword{internal}
