% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{SDALGCPexceedanceST}
\alias{SDALGCPexceedanceST}
\title{Exceedance probability of the relative risk}
\usage{
SDALGCPexceedanceST(obj, thresholds, continuous = TRUE)
}
\arguments{
\item{obj}{an object of class "Pred.SDALGCPST" obtained as result of a call to \code{\link{SDALGCPPred_ST}}.}

\item{thresholds}{either a vector  of numbers or a vector of single value.}

\item{continuous}{logical; TRUE for spatially continuous relative risk and FALSE for region specific relative risk. default is TRUE}
}
\value{
A vector or dataframe(for more than one value of the threshold) of the exceedance probability
}
\description{
Computes the exceedance probability for a given threshold of the spatially continuous relative risk or the region specific relative risk from the object of class "Pred.SDALGCP".
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
\keyword{internal}
