% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{SDALGCPPred}
\alias{SDALGCPPred}
\title{Prediction from Fitted SDA-LGCP Model}
\usage{
SDALGCPPred(para_est, cellsize = NULL, continuous = TRUE, bound = NULL)
}
\arguments{
\item{para_est}{Output object from \code{\link{SDALGCPMCML}}, of class `"SDALGCP"`.}

\item{cellsize}{Grid resolution (in projection units) for continuous prediction.}

\item{continuous}{Logical; if `TRUE`, performs spatially continuous prediction. If `FALSE`, performs region-specific discrete prediction.}

\item{bound}{Optional `sf` object representing the prediction boundary (used for continuous prediction). Defaults to the boundary of the model.}
}
\value{
An object of class `"Pred.SDALGCP"` containing predictions, standard errors, and coordinates.
}
\description{
Delivers spatially discrete or continuous prediction from the output of the `SDALGCPMCML` model.
}
\examples{
# check vignette for examples
}
