% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{SDALGCPUniformPoint}
\alias{SDALGCPUniformPoint}
\title{Generate Uniform Random Points in a Polygon}
\usage{
SDALGCPUniformPoint(
  poly,
  delta,
  weighted = FALSE,
  pop = NULL,
  pop_shp = NULL,
  lambdamax = NULL,
  n = NULL,
  rho = NULL,
  giveup = NULL,
  bound
)
}
\arguments{
\item{poly}{matrix of coordinates (not used directly when `bound` is provided).}

\item{delta}{unused in this method (included for consistency).}

\item{weighted}{logical; whether to use population-weighted sampling. Default is FALSE.}

\item{pop}{numeric vector of population weights (if weighted = TRUE).}

\item{pop_shp}{optional; an `sf` object representing population distribution polygons.}

\item{lambdamax}{optional; maximum population density value (used only if `weighted = TRUE`).}

\item{n}{optional; number of points to generate. If NULL, will be inferred using `rho`.}

\item{rho}{optional; intensity (points per unit area). Used to determine `n` if not supplied.}

\item{giveup}{optional; max attempts for rejection sampling (used if `weighted = TRUE`).}

\item{bound}{an `sf` polygon object representing the sampling boundary.}
}
\value{
A list with:
\describe{
  \item{xy}{matrix of sampled coordinates.}
  \item{win}{spatstat window object used.}
}
}
\description{
Uniformly generates random points within a polygon defined by an `sf` boundary.
}
