% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{SDAProfilePhi}
\alias{SDAProfilePhi}
\title{Profile Likelihood of Phi}
\usage{
SDAProfilePhi(obj)
}
\arguments{
\item{obj}{an object of class "SDALGCP" from \code{\link{SDALGCPMCML}}.}
}
\value{
A base R plot.
}
\description{
Plots the log-likelihood profile of the spatial scale parameter \code{phi}.
}
