% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{summary.SDALGCP}
\alias{summary.SDALGCP}
\title{Summary method for SDALGCP model object}
\usage{
\method{summary}{SDALGCP}(object, ...)
}
\arguments{
\item{object}{An object of class \code{SDALGCP} obtained from \code{\link{SDALGCPMCML}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list with components:
\describe{
\item{parameter_estimate_result}{Matrix of parameter estimates, standard errors, z-values and p-values.}
\item{phi}{Scale parameter estimate.}
\item{ll}{Value of the log-likelihood at maximum.}
\item{call}{Matched function call.}
}
}
\description{
Computes standard errors and p-values for parameter estimates from an SDALGCP model.
}
