% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_functions.R
\name{summary.SDALGCPST}
\alias{summary.SDALGCPST}
\title{Summarizing the parameter estimates of SDALGCP model}
\usage{
\method{summary}{SDALGCPST}(object, ...)
}
\arguments{
\item{object}{an object of class "SDALGCPST" obtained as result of a call to \code{\link{SDALGCPMCML}} .}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list with the following components

\code{parameter_estimate_result}: the parameter of the SDALGCP model

\code{phi}: the scale parameter of the Gaussian process

\code{ll}: value of likelihood function at the maximum likelihood estimates.

\code{call}: matched call.
}
\description{
\code{summary} method for the class "SDALGCPST" that computes the standard errors and p-values of SDALGCPST.
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
