% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{confint.SDALGCP}
\alias{confint.SDALGCP}
\title{Confidence Intervals for SDALGCP Model Parameters}
\usage{
\method{confint}{SDALGCP}(object, parm, level = 0.95, dp = 3, ...)
}
\arguments{
\item{object}{an object of class "SDALGCP" obtained as result of a call to \code{\link{SDALGCPMCML}}.}

\item{parm}{a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{dp}{the number of decimal places for the result}

\item{...}{additional argument(s) for methods.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in % (by default 2.5% and 97.5%).
}
\description{
Computes confidence intervals for one or more parameters in a fitted SDALGCP model from the object of class "SDALGCP", based on asymptotic normality.
}
\examples{
\dontrun{
data(PBCshp)
data <- as.data.frame(PBCshp@data)
FORM <- X ~ propmale + Income + Employment + Education + Barriers + Crime + 
Environment +  offset(log(pop))
phi <- seq(500, 1700, length.out = 20)
control.mcmc <- list(n.sim = 110000, burnin = 10000, thin= 10, h=1.65/(545^(1/6)),
                    c1.h = 0.01, c2.h = 1e-04)
my_est <- SDALGCPMCML(data=data, formula=FORM, my_shp=PBCshp, delta=100, phi=phi, method=1, 
                     weighted=FALSE,  plot=TRUE, par0=NULL, control.mcmc=control.mcmc)
confint(my_est)
}
}
\seealso{
\link{confint.lm}, \link{confint.default}, \link{SDALGCPMCML}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
