% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_functions.R
\name{plot_continuous}
\alias{plot_continuous}
\title{Plot of the spatially continuous prediction over the entire region}
\usage{
plot_continuous(obj, bound = NULL, type = "relrisk", ...)
}
\arguments{
\item{obj}{an object of class "Pred.SDALGCP" obtained as result of a call to \code{\link{SDALGCPPred}}}

\item{bound}{the boundary of the predictive grid, not required if predictive grid is not supplied}

\item{type}{Character string: what type of plot to produce. Choices are "relrisk" (=exp(S)); "SErelrisk" (standard error of the relative risk).}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\value{
The function does not return any value.
}
\description{
This function plots the region specific incidence and covariate adjusted relative risk and their corresponding standard error using the output from \code{\link{SDALGCPPred}} function.
}
\seealso{
\code{\link{SDALGCPPred}}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
