% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{plot_continuous}
\alias{plot_continuous}
\title{plot_continuous}
\usage{
plot_continuous(obj, bound = NULL, type = "relrisk", overlay = FALSE, ...)
}
\arguments{
\item{obj}{an object of class "Pred.SDALGCP" obtained as result of a call to \code{\link{SDALGCPPred}}}

\item{bound}{the boundary of the predictive grid, not required if predictive grid is not supplied}

\item{type}{Character string: what type of plot to produce. Choices are "relrisk" (=exp(S)); "SErelrisk" (standard error of the relative risk).}

\item{overlay}{optional; a logical operation to indicate either to add a base map.}

\item{...}{further arguments passed to \code{\link{plot}}.}
}
\value{
The function does not return any value.
}
\description{
A generic function for mapping spatially continuous prediction for \code{\link{SDALGCPPred}} function in \link{SDALGCP} package. Not for general purposes
}
\seealso{
\code{\link{SDALGCPPred}}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
\keyword{internal}
