% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{plot_SDALGCPexceedance}
\alias{plot_SDALGCPexceedance}
\title{plot_SDALGCPexceedance}
\usage{
plot_SDALGCPexceedance(
  obj,
  thresholds,
  bound = NULL,
  continuous = TRUE,
  overlay = FALSE,
  ...
)
}
\arguments{
\item{obj}{an object of class "Pred.SDALGCP" obtained as result of a call to \code{\link{SDALGCPPred}}.}

\item{thresholds}{either a vector  of numbers or a vector of single value.}

\item{bound}{optional; it gives the boundary of the region, only useful when the predictive location is supplied in \link{SDALGCPPred}.}

\item{continuous}{logical; TRUE for spatially continuous relative risk and FALSE for region specific relative risk. default is TRUE.}

\item{overlay}{optional; a logical operation to indicate either to add a base map.}

\item{...}{further arguments passed to \link{plot}.}
}
\value{
The plot of the exceedance probability map
}
\description{
A generic function for mapping the exceedance probability for a given threshold of the spatially continuous relative risk or the region specific relative risk from the object of class "Pred.SDALGCP". Not for general purposes.
}
\seealso{
\link{SDALGCPPred}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
\keyword{internal}
