% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_den.R
\docType{data}
\name{pop_den}
\alias{pop_den}
\title{Population density raster shapefile of Newcastle Upon Tyne in the UK.}
\format{This is a raster file of population density, 300 X 300 meters.}
\usage{
data(pop_den)
}
\description{
A dataset containing the population density of Newcastle upon Tyne of 300 X 300 metres
}
\references{
Murdock, A.P., Harfoot, A.J.P., Martin, D., Cockings, S. and Hill, C. (2015) OpenPopGrid: an open gridded population dataset for England and Wales. GeoData, University of Southampton.
}
\keyword{datasets}
