% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_functions.R
\name{precomputeCorrMatrix}
\alias{precomputeCorrMatrix}
\title{Pre-compute the covariance matrix}
\usage{
precomputeCorrMatrix(S.coord, phi)
}
\arguments{
\item{S.coord}{The list of the coordinates of the points created in each polygon}

\item{phi}{The vector of the scale parameter}
}
\value{
return an array of the covariance matrix for the specified vectors of phi
}
\description{
This function precomputes the covariance matrix
}
\details{
This function precompute the covariance matrix using the exponential covariance function
}
\examples{
\donttest{
load(popshape.RData)
xy <- gridPointsSSIP(shp=popshape, delta, rho=0.55, giveup=1000, divisor=1)
pre.matrix <- precompute.corr.matrix(S.coord=xy, phi=c(1:5))
}
}
\seealso{
\code{\link{SDALGCPpolygonpoints}},
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
