% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_functions.R
\name{SDALGCPMCML}
\alias{SDALGCPMCML}
\title{Parameter estimation for SDA-LGCP Using Monte Carlo Maximum likelihood}
\usage{
SDALGCPMCML(data, formula, my_shp, delta, phi = NULL, method = 1,
  pop_shp = NULL, weighted = FALSE, plot = TRUE, par0 = NULL,
  control.mcmc = NULL, plot_profile = FALSE, rho = NULL)
}
\arguments{
\item{data}{data frame containing the variables in the model.}

\item{formula}{an object of class \code{\link{formula}} (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{my_shp}{A SpatialPolygons orSpatialPolygonsDataFrame  object containing the polygons (i.e each regions).}

\item{delta}{distance between points}

\item{phi}{the discretised values of the scale parameter phi. if not supplied, it uses the default.}

\item{method}{To specify which method to use to sample the points, the options are 1 for Simple Sequential Inhibition (SSI) process, 2 for Uniform sampling and 3 for regular grid. 1 is the default}

\item{pop_shp}{Optional, The raster of population density map for population weighted approach}

\item{weighted}{To specify if you want to use the population density, default to FALSE, i.e population density is not used.}

\item{plot}{To display the plot of the points inside the polygon, default to TRUE}

\item{par0}{the initial parameter of the fixed effects beta and the variance sigmasq, specified as c(beta, sigma2)}

\item{control.mcmc}{list from PrevMap package to define the burnin, thining, the number of iteration and the turning parameters see \code{\link{controlmcmcSDA}}.}

\item{plot_profile}{logical; if TRUE the profile-likelihood is plotted. default is FALSE}

\item{rho}{Optional, The packing density, default set to 0.55}
}
\value{
An object of class "PrevMap".
The function \code{\link{summary.PrevMap}} is used to print a summary of the fitted model.
The object is a list with the following components:

\code{D}: matrix of covariates.

\code{y}: the count, response observations.

\code{m}: offset

\code{beta_opt}: estimates of the fixed effects of the model.

\code{sigma2_opt}: estimates of the variance of the Gaussian process.

\code{phi_opt}: estimates of the scale parameter phi of the Gaussian process.

\code{cov}: covariance matrix of the MCML estimates.

\code{Sigma_mat_opt}: covariance matrix of the Gaussian process that corresponds to the optimal value

\code{llike_val_opt}: maximum value of the log-likelihood.

\code{mu}: mean of the linear predictor

\code{all_para}: the entire estimates for the different values of phi.

\code{all_cov}: the entire covariance matrix of the estimates for the different values of phi.

\code{par0}: the initial parameter of the fixed effects beta and the variance sigmasq used in the estimation

\code{control.mcmc}: the burnin, thining, the number of iteration and the turning parameters used see \code{\link{controlmcmcSDA}}.

\code{call}: the matched call.
}
\description{
This function provides the maximum likelihood estimation of the parameter given the precomputed correlation matrices for different values of scale parameter, phi
}
\details{
This function performs parameter estimation for a SDA-LGCP Model
\bold{Monte Carlo Maximum likelihood.}
The Monte Carlo maximum likelihood method uses conditional simulation from the distribution of the random effect \eqn{T(x) = d(x)'\beta+S(x)} given the data \code{y}, in order to approximate the high-dimensiional intractable integral given by the likelihood function. The resulting approximation of the likelihood is then maximized by a numerical optimization algorithm which uses analytic epression for computation of the gradient vector and Hessian matrix. The functions used for numerical optimization are \code{\link{nlminb}}
}
\examples{
\donttest{
data <- as.data.frame(PBCshp@data)
FORM <- X ~ propmale + Income + Employment + Education + Barriers + Crime + 
Environment +  offset(log(pop))
phi <- seq(500, 1700, length.out = 20)
control.mcmc <- list(n.sim = 110000, burnin = 10000, thin= 10, h=1.65/(545^(1/6)),
                    c1.h = 0.01, c2.h = 1e-04)
my_est <- SDALGCPMCML(data=data, formula=FORM, my_shp=PBCshp, delta=100, phi=phi, method=1, 
                     weighted=FALSE,  plot=TRUE, par0=NULL, control.mcmc=control.mcmc)
}
}
\references{
Giorgi, E., & Diggle, P. J. (2017). PrevMap: an R package for prevalence mapping. Journal of Statistical Software, 78(8), 1-29. doi:10.18637/jss.v078.i08

Christensen, O. F. (2004). Monte carlo maximum likelihood in model-based geostatistics. Journal of Computational and Graphical Statistics 13, 702-718.
}
\seealso{
\link{Aggregated_poisson_log_MCML}, \code{\link{Laplace.sampling}}
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
