% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{SDALGCPpolygonpoints}
\alias{SDALGCPpolygonpoints}
\title{Generate Sampling Points from a Polygon Layer}
\usage{
SDALGCPpolygonpoints(
  my_shp,
  delta,
  method = 1,
  pop_shp = NULL,
  weighted = FALSE,
  rho = NULL,
  plot = FALSE,
  giveup = NULL
)
}
\arguments{
\item{my_shp}{An `sf` object containing polygon geometries.}

\item{delta}{Grid resolution or minimum separation distance between candidate points.}

\item{method}{Integer specifying the sampling method: 1 = SSI, 2 = Uniform, 3 = Regular.}

\item{pop_shp}{Optional `terra::rast` object representing population density.}

\item{weighted}{Logical indicating whether to use population weighting.}

\item{rho}{Optional packing density for SSI (default = 0.55).}

\item{plot}{Logical. If TRUE, plots the generated points.}

\item{giveup}{Optional integer for maximum failed SSI attempts (default = 1000).}
}
\value{
A list of length equal to number of polygons in `my_shp`. Each element contains:
\describe{
  \item{xy}{matrix of candidate point coordinates}
  \item{n}{number of points}
  \item{bound}{the polygon geometry used}
  }
}
\description{
Creates candidate sampling points within each polygon feature of an `sf` object, either uniformly or weighted by population density.
}
\examples{
\donttest{
data(PBCshp)
PBCsf <- sf::st_as_sf(PBCshp)
pts <- SDALGCPpolygonpoints(my_shp = PBCsf, delta = 500, method = 3)
head(pts[[1]]$xy)
}
}
