% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_functions.R
\name{SDALGCPPred}
\alias{SDALGCPPred}
\title{Spatial continuous predictions of the relative risk using plug-in of MCML estimates}
\usage{
SDALGCPPred(para_est, cellsize, continuous = TRUE, control.mcmc = NULL,
  pred.loc = NULL, divisor = 1, plot.correlogram = F, messages = TRUE,
  parallel = FALSE)
}
\arguments{
\item{para_est}{an object of class "SDALGCP" obtained as a result of a call to \code{\link{SDALGCPMCML}}.}

\item{cellsize}{the size of the computational grid}

\item{continuous}{logical; to choose which prrediction to do perform, discrete or continuous. the default is continuous.}

\item{control.mcmc}{output from \code{\link{controlmcmcSDA}}, if not provided, it uses the values used for the parameter estimation}

\item{pred.loc}{optional, the dataframe of the predictive grid.}

\item{divisor}{optional, the value to use to convert the dimension of the polygon, default is 1 which implies no conversion}

\item{plot.correlogram}{ogical; if plot.correlogram=TRUE the autocorrelation plot of the conditional simulations is displayed.}

\item{messages}{logical; if messages=TRUE then status messages are printed on the screen (or output device) while the function is running. Default is messages=TRUE.}

\item{parallel}{to parallelize some part of the function.}
}
\value{
pred.draw: the samples of the prediction

pred: the prediction of the relative risk

predSD: the standard error of the prediction

Pred.loc: The cordinates of the predictive locations
}
\description{
This function performs spatial continuous prediction, fixing the model parameters at the Monte Carlo maximum likelihood estimates of a SDA-LGCP model.
}
\details{
The function returns the prediction of the relative risk exp(S(x))
}
\examples{
\donttest{
load("PBCshp")
data <- as.data.frame(PBCshp@data)
FORM <- X ~ propmale + Income + Employment + Education + Barriers + Crime + 
Environment +  offset(log(pop))
phi <- seq(500, 1700, length.out = 20)
control.mcmc <- list(n.sim = 110000, burnin = 10000, thin= 10, h=1.65/(545^(1/6)),
                    c1.h = 0.01, c2.h = 1e-04)
my_est <- SDALGCPMCML(data=data, formula=FORM, my_shp=PBCshp, delta=100, phi=phi, method=1, 
                     weighted=FALSE,  plot=TRUE, par0=NULL, control.mcmc=control.mcmc)
Con_pred <- SDALGCPPred(para_est=my_est,  cellsize=300, continuous=TRUE)
}
}
\references{
Banerjee, S., Carlin, B. P., & Gelfand, A. E. (2014). Hierarchical modeling and analysis for spatial data. CRC press.
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
