% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_functions.R
\name{SDAProfilePhi}
\alias{SDAProfilePhi}
\title{plot profile likelihood of phi}
\usage{
SDAProfilePhi(obj)
}
\arguments{
\item{obj}{the output of \code{\link{SDALGCPMCML}} of class "SDALGCP"}
}
\value{
A plot
}
\description{
This function plots the profile likelihood of phi
}
\details{
To be used to view the value of the likelihood versus the scale parameter phi
}
\author{
Olatunji O. Johnson \email{o.johnson@lancaster.ac.uk}

Emanuele Giorgi \email{e.giorgi@lancaster.ac.uk}

Peter J. Diggle \email{p.diggle@lancaster.ac.uk}
}
